requirements.txt
# Dropout Prediction System - Requirements

# Core ML Libraries
catboost==1.2.5
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Database
mysql-connector-python==8.2.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Visualization (for future dashboard)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# API Framework (for future backend)
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Additional utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
